requirements.txt
blinker==1.6.2
certifi==2023.5.7
click==8.1.3
email_validator==2.0.0.post2
Flask==2.3.2
flask_behind_proxy==0.1.1
flask_sqlalchemy==3.0.5
Flask-WTF==1.1.1
GitPython==3.1.31
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycodestyle
pytest==7.4.0
Werkzeug==2.3.6
google-search-results==2.4.2
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
dataclasses-json==0.5.9
dnspython==2.3.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
requests==2.31.0
requests-oauthlib==1.3.1
python-youtube==0.9.0
urllib3==2.0.3
gitdb==4.0.10
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
packaging==23.1
pip==23.1.2
pluggy==1.2.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.3
smmap==5.0.0
setuptools==65.6.3
countryinfo==0.1.2
psycopg2-binary==2.9.6